# openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: RESTful EPP
  version: "1.0"
servers:
- url: http://localhost:8080/repp/v1/
paths:
...
  /domains:
    post:
      operationId: domains_post
      parameters:
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainCreateRequest'
        required: true
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Domain create
      tags:
      - domains
      x-openapi-router-controller: openapi_server.controllers.domains_controller
  /domains/{id}:
    delete:
      operationId: domains_id_delete
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Domain delete
      tags:
      - domains
      x-openapi-router-controller: openapi_server.controllers.domains_controller
    get:
      operationId: domains_id_get
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      - description: Object authorization details
        explode: false
        in: header
        name: REPP-authInfo
        required: false
        schema:
          type: string
        style: simple
      - description: Object linked to authorization
        explode: false
        in: header
        name: REPP-Roid
        required: false
        schema:
          type: string
        style: simple
      - description: Name of attibute to filter on
        explode: true
        in: query
        name: filter
        required: false
        schema:
          type: string
        style: form
      - description: Value to use for filter
        explode: true
        in: query
        name: val
        required: false
        schema:
          type: string
        style: form
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "422":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Domain info
      tags:
      - domains
      x-openapi-router-controller: openapi_server.controllers.domains_controller
    head:
      operationId: domains_id_head
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          description: OK
          headers:
            REPP-Check-Avail:
              $ref: '#/components/headers/check-avail'
            REPP-Check-Reason:
              $ref: '#/components/headers/check-reason'
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "422":
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Domain check
      tags:
      - domains
      x-openapi-router-controller: openapi_server.controllers.domains_controller
    patch:
      operationId: domains_id_patch
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Domain update
      tags:
      - domains
      x-openapi-router-controller: openapi_server.controllers.domains_controller
  /domains/{id}/renewals:
    post:
      operationId: domains_id_renewals_post
      parameters:
      - description: Object identifier
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      - description: Unit used for renewal value (e.g. y for year)
        example: "y"
        explode: true
        in: query
        name: unit
        required: false
        schema:
          type: string
        style: form
      - description: Value for renewal
        example: 1
        explode: true
        in: query
        name: value
        required: false
        schema:
          type: integer
        style: form
      - description: Date on which the current validity period ends
        example: 2023-12-01
        explode: true
        in: query
        name: current-date
        required: false
        schema:
          type: string
        style: form
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Domain renew
      tags:
      - domains
      x-openapi-router-controller: openapi_server.controllers.domains_controller
...
components:
  headers:
    check-avail:
      description: Result of object check
      explode: false
      required: true
      schema:
        type: integer
      style: simple
    check-reason:
      description: Explanation why object cannot be provisioned
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    cltrid:
      description: Client transaction identifier
      explode: false
      required: false
      schema:
        type: string
      style: simple
    eppcode:
      description: EPP result code
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    queue-size:
      description: Number of messages in the queue
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    svtrid:
      description: Server transaction identifier
      explode: false
      required: false
      schema:
        type: string
      style: simple
...
  requestBodies:
    defaultOptionalRequestBody:
      content:
        application/epp+json:
          schema:
            type: string
        application/epp+xml:
          schema:
            type: string
      description: Default request body for XML of JSON message
      required: false
    defaultRequestBody:
      content:
        application/epp+json:
          schema:
            type: string
        application/epp+xml:
          schema:
            type: string
      description: Default request body for XML of JSON message
      required: true
...
  schemas:
    DomainCreateRequest:
      allOf:
      - $ref: '#/components/schemas/epp_eppType'
      - properties:
          command:
            $ref: '#/components/schemas/DomainCreateRequest_allOf_command'
        required:
        - command
      example:
        command:
          extension:
            secDNS:
              maxSigLife: 1294386359
          create:
            period:
              '#value': 8
              '@unit': "y"
            ns:
              hostObj:
              - null
              - null
            contact:
            - '@type': admin
              '#value': '#value'
            - '@type': admin
              '#value': '#value'
            name: name
            registrant: registrant
            authInfo:
              pw:
                '@roid': '@roid'
                '#value': '#value'
      title: DomainCreateRequest
...
    epp_eppType:
      oneOf:
      - $ref: '#/components/schemas/epp_eppType_oneOf'
      - $ref: '#/components/schemas/epp_eppType_oneOf_1'
      - $ref: '#/components/schemas/epp_eppType_oneOf_2'
      - $ref: '#/components/schemas/epp_eppType_oneOf_3'
      - $ref: '#/components/schemas/epp_eppType_oneOf_4'
      title: epp_eppType
...
    epp_greetingType:
      properties:
        dcp:
          $ref: '#/components/schemas/epp_dcpType'
        svDate:
          format: date-time
          title: svDate
          type: string
        svID:
          maxLength: 64
          minLength: 3
          title: epp_sIDType
          type: string
        svcMenu:
          $ref: '#/components/schemas/epp_svcMenuType'
      required:
      - dcp
      - svDate
      - svID
      - svcMenu
      title: epp_greetingType
      type: object
...
    DomainCreateRequest_allOf_command_extension:
      example:
        secDNS:
          maxSigLife: 1294386359
      properties:
        secDNS:
          $ref: '#/components/schemas/secDNS_dsOrKeyType'
      title: DomainCreateRequest_allOf_command_extension
      type: object
    DomainCreateRequest_allOf_command:
      example:
        extension:
          secDNS:
            maxSigLife: 1294386359
        create:
          period:
            '#value': 8
            '@unit': "y"
          ns:
            hostObj:
            - null
            - null
          contact:
          - '@type': admin
            '#value': '#value'
          - '@type': admin
            '#value': '#value'
          name: name
          registrant: registrant
          authInfo:
            pw:
              '@roid': '@roid'
              '#value': '#value'
      properties:
        create:
          $ref: '#/components/schemas/domain_createType'
        extension:
          $ref: '#/components/schemas/DomainCreateRequest_allOf_command_extension'
      required:
      - create
      title: DomainCreateRequest_allOf_command
      type: object
...
    epp_eppType_oneOf:
      properties:
        greeting:
          $ref: '#/components/schemas/epp_greetingType'
      required:
      - greeting
      title: epp_eppType_oneOf
      type: object
    epp_eppType_oneOf_1:
      properties:
        hello:
          default: false
          title: hello
          type: boolean
      required:
      - hello
      title: epp_eppType_oneOf_1
      type: object
    epp_eppType_oneOf_2:
      properties:
        command:
          $ref: '#/components/schemas/epp_commandType'
      required:
      - command
      title: epp_eppType_oneOf_2
      type: object
    epp_eppType_oneOf_3:
      properties:
        response:
          $ref: '#/components/schemas/epp_responseType'
      required:
      - response
      title: epp_eppType_oneOf_3
      type: object
    epp_eppType_oneOf_4:
      properties:
        extension:
          $ref: '#/components/schemas/epp_extAnyType'
      required:
      - extension
      title: epp_eppType_oneOf_4
      type: object
...